fix(core)!: do not automatically fetch generic URLs
Do not automatically fetch URLs, unless configured to do so. To continue
to automatically fetch and evaluate the response of URLs, install and
load the new URL plugin like so:

    require('@citation-js/plugin-url')

This is a breaking change. This change does not apply to specific URLs,
like the DOI, ISBN, Pubmed, and Wikidata plugins. The @else/url input
type can still be used to build input types off of.